Add files via upload
this app shows theoperation of a traditional cryptographic system. Inparticular, this program is a variant of the Enigma coding machine that the German army used
from the beginning of the thirties until the end of the second war world.
This is the executable version to a desktop app, me and my team have designed, using the tkinter GUI on python. you can check the code on the main.py file, check the repository, to find out about the utility of the enigma machine.
you can check this link, to understand how is the enigma machine concieved : https://en.wikipedia.org/wiki/Enigma_machine